+++ lib/src/annotations/annotations.dart
@@ -11,6 +11,8 @@ export 'cache/cache_annotation.dart';
 export 'configuration/configuration_annotations.dart';
 export 'form/form_annotations.dart';
 export 'http/http_annotations.dart';
+export 'http/streaming_annotation.dart';
+
 export 'models/rest_model.dart';
 export 'multipart/multipart_annotations.dart';
 export 'parameters/parameter_annotations.dart';